fix: change Focuser.stepsize return type from int to float (#16)
The ASCOM specification defines stepsize as a double (float in Python)
since it represents microns which can be fractional values.

Fixes #9

+++ python_alpaca_server/api/focuser.py
@@ -68,8 +68,8 @@ def create_router(devices: List[Device]):
     async def get_stepsize(
         req: Annotated[CommonRequest, Query()],
         device: Focuser = Depends(device_finder(devices, UrlDeviceType.Focuser)),
-    ) -> Response[int]:
-        return Response[int].from_request(
+    ) -> Response[float]:
+        return Response[float].from_request(
             req,
             device.get_stepsize(req),
         )
